Change gradle file
