+++ flank-scripts/build.gradle.kts
@@ -99,7 +99,6 @@ publishing {
 tasks.test {
     maxHeapSize = "2048m"
     minHeapSize = "512m"
-    systemProperty("testScript", "")
 }
 
 repositories {
