+++ flank-scripts/src/main/kotlin/flank/scripts/github/GithubApi.kt
@@ -14,9 +14,9 @@ import com.jcabi.github.Release
 import com.jcabi.github.Releases
 import com.jcabi.github.Repo
 import com.jcabi.github.RtGithub
+import flank.common.config.flankRepository
 import flank.scripts.ci.releasenotes.GitHubRelease
 import flank.scripts.ci.releasenotes.GithubReleaseDeserializable
-import flank.scripts.config.flankRepository
 import flank.scripts.exceptions.mapClientErrorToGithubException
 import flank.scripts.github.objects.GitHubCommit
 import flank.scripts.github.objects.GitHubCommitListDeserializer
