+++ flank-scripts/src/main/kotlin/flank/scripts/integration/WorkflowSummary.kt
@@ -1,6 +1,6 @@
 package flank.scripts.integration
 
-import flank.scripts.config.fullSuiteWorkflowFilename
+import flank.common.config.fullSuiteWorkflowFilename
 import flank.scripts.github.commons.getLastWorkflowRunDate
 
 suspend fun getLastITWorkflowRunDate(token: String) = getLastWorkflowRunDate(
