+++ flank-scripts/src/main/kotlin/flank/scripts/shell/firebase/sdk/LastSDKUpdateRun.kt
@@ -1,6 +1,6 @@
 package flank.scripts.shell.firebase.sdk
 
-import flank.scripts.config.updateDependenciesWorkflowFilename
+import flank.common.config.updateDependenciesWorkflowFilename
 import flank.scripts.github.commons.getLastWorkflowRunDate
 
 suspend fun getLastSDKUpdateRunDate(token: String) = getLastWorkflowRunDate(
